Acceptable Json For Entities
This json is a bit of a modification to create/update objects for the application. There were crashing errors that was stopping me from completing the project that were difficult to pinpoint but this json is sufficient in carrying duties. All other methods are able to be completed using methods described in Functional Testing Instructions.